feat: refactor article single page (#214)
