+++ app/layout.tsx
@@ -8,7 +8,7 @@ export default function RootLayout({
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
+    <html lang="en" className="h-full antialiased">
       <head>
         <title>Vitalii - Software engineer, builder, and amateur runner.</title>
         <meta
@@ -16,17 +16,19 @@ export default function RootLayout({
           content="I’m Vitalii, a software engineer  based in Porto, Portugal. I occasinally build new stuff and share my learnings here."
         />
       </head>
-      <body>
-        <div className="fixed inset-0 flex justify-center">
-          <div className="flex w-full max-w-7xl">
-            <div className="w-full bg-white ring-1 ring-zinc-100 dark:bg-zinc-900 dark:ring-zinc-300/20" />
+      <body className="flex h-full flex-col bg-zinc-50 dark:bg-black">
+        <>
+          <div className="fixed inset-0 flex justify-center">
+            <div className="flex w-full max-w-7xl">
+              <div className="w-full bg-white ring-1 ring-zinc-100 dark:bg-zinc-900 dark:ring-zinc-300/20" />
+            </div>
           </div>
-        </div>
-        <div className="relative">
-          <Header />
-          <main>{children}</main>
-          <Footer />
-        </div>
+          <div className="relative">
+            <Header />
+            <main>{children}</main>
+            <Footer />
+          </div>
+        </>
       </body>
     </html>
   )
