+++ package.json
@@ -32,14 +32,12 @@
   "dependencies": {
     "@headlessui/react": "^1.7.4",
     "@heroicons/react": "^2.0.13",
+    "date-fns": "^2.29.3",
     "next": "^13.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "clsx": "^1.2.1",
-    "typescript": "^4.5.5",
-    "next-mdx-remote": "^4.3.0",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/typography": "^0.5.8",
@@ -48,6 +46,7 @@
     "@typescript-eslint/eslint-plugin": "^5.45.0",
     "@typescript-eslint/parser": "^5.45.0",
     "autoprefixer": "^10.4.13",
+    "clsx": "^1.2.1",
     "eslint": "8.26.0",
     "eslint-config-next": "13.0.2",
     "eslint-config-prettier": "^8.5.0",
@@ -55,9 +54,11 @@
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-react": "^7.29.2",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "next-mdx-remote": "^4.3.0",
     "postcss": "^8.4.19",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.11",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.5.5"
   }
 }
