+++ src/components/ArticleCard.tsx
@@ -1,16 +1,6 @@
-'use client'
-import Link from 'next/link'
-
 import { ArticleMarkdownItem } from '../lib/markdown'
-
-export function formatDate(date: Date) {
-  return date.toLocaleString('en-US', {
-    day: 'numeric',
-    month: 'long',
-    year: 'numeric',
-    timeZone: 'UTC',
-  })
-}
+import { formatDate } from '../lib/date'
+import { ArticleCardTitle } from './ArticleCardTitle'
 
 function Eyebrow({ date }: { date: Date }) {
   return (
@@ -48,15 +38,7 @@ export function ArticleCard({ article }: { article: ArticleMarkdownItem }) {
   return (
     <article className="group relative flex flex-col items-start">
       <>
-        <h2 className="text-base font-semibold tracking-tight text-zinc-800 dark:text-zinc-100">
-          <>
-            <div className="absolute -inset-y-6 -inset-x-4 z-0 scale-95 bg-zinc-50 opacity-0 transition group-hover:scale-100 group-hover:opacity-100 dark:bg-zinc-800/50 sm:-inset-x-6 sm:rounded-2xl" />
-            <Link href={`/articles/${article.slug}`}>
-              <span className="absolute -inset-y-6 -inset-x-4 z-20 sm:-inset-x-6 sm:rounded-2xl" />
-              <span className="relative z-10">{article.meta.title}</span>
-            </Link>
-          </>
-        </h2>
+        <ArticleCardTitle url={article.url} title={article.meta.title} />
         <Eyebrow date={article.meta.date} />
         <p className="relative z-10 mt-2 text-sm text-zinc-600 dark:text-zinc-400">
           {article.meta.description}
