+++ src/lib/date.ts
@@ -1,0 +1,5 @@
+import { format } from 'date-fns'
+
+export function formatDate(date: Date) {
+  return format(date, 'LLLL d, yyyy')
+}
