--- utils/date.ts
@@ -1,19 +1,0 @@
-const shortMonthNames = [
-  'Jan',
-  'Feb',
-  'Mar',
-  'Apr',
-  'May',
-  'Jun',
-  'Jul',
-  'Aug',
-  'Sep',
-  'Oct',
-  'Nov',
-  'Dec',
-]
-
-const getMonthName = (date: Date): string => shortMonthNames[date.getMonth()]
-
-export const getFormattedDate = (date: Date): string =>
-  `${date.getDate()} ${getMonthName(date)} ${date.getFullYear()}`
