--- utils/media.ts
@@ -1,2 +1,0 @@
-export const MOBILE_MAX_WITH = '480px'
-export const TABLET_MAX_WITH = '768px'
